Tweaks to UI and components for better UX

+++ core/api/promotionReservations/server/methods.js
@@ -4,44 +4,28 @@ import {
   promotionReservationInsert,
   promotionReservationRemove,
   promotionReservationUpdate,
+  promotionReservationUpdateObject,
 } from '../methodDefinitions';
 
 promotionReservationInsert.setHandler(({ userId }, params) => {
-  const { promotionOptionId } = params;
-  SecurityService.promotions.isAllowedToManagePromotionReservation({
-    promotionOptionId,
-    userId,
-  });
+  SecurityService.checkUserIsAdmin(userId);
   return PromotionReservationService.insert(params);
 });
 
 promotionReservationRemove.setHandler(({ userId }, { promotionReservationId }) => {
-  const {
-    promotionOption: { _id: promotionOptionId },
-  } = PromotionReservationService.safeFetchOne({
-    $filters: { _id: promotionReservationId },
-    promotionOption: { _id: 1 },
-  });
-  SecurityService.promotions.isAllowedToManagePromotionReservation({
-    promotionOptionId,
-    userId,
-  });
-  return PromotionReservationService.remove(promotionReservationId);
+  SecurityService.checkUserIsAdmin(userId);
+  PromotionReservationService.remove(promotionReservationId);
 });
 
 promotionReservationUpdate.setHandler(({ userId }, { promotionReservationId, object }) => {
-  const {
-    promotionOption: { _id: promotionOptionId },
-  } = PromotionReservationService.safeFetchOne({
-    $filters: { _id: promotionReservationId },
-    promotionOption: { _id: 1 },
-  });
-  SecurityService.promotions.isAllowedToManagePromotionReservation({
-    promotionOptionId,
-    userId,
-  });
+  SecurityService.checkUserIsAdmin(userId);
   return PromotionReservationService._update({
     id: promotionReservationId,
     object,
   });
+});
+
+promotionReservationUpdateObject.setHandler(({ userId }, params) => {
+  SecurityService.checkUserIsAdmin(userId);
+  return PromotionReservationService.updateStatusObject(params);
 });
